requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
pydantic-settings
python-dotenv
passlib[bcrypt]
